+++ mimikatz/modules/crypto/kuhl_m_crypto_extractor.h
@@ -7,6 +7,7 @@
 #include "../kuhl_m.h"
 #include "../../../modules/kull_m_crypto.h"
 #include "../../../modules/kull_m_memory.h"
+#include "../../../modules/kull_m_process.h"
 
 void kuhl_m_crypto_extractor_capi32(PKULL_M_MEMORY_ADDRESS address);
 void kuhl_m_crypto_extractor_bcrypt32(PKULL_M_MEMORY_ADDRESS address);
@@ -17,6 +18,8 @@ void kuhl_m_crypto_extractor_bcrypt64(PKULL_M_MEMORY_ADDRESS address);
 
 DWORD kuhl_m_crypto_extractor_GetKeySizeForEncryptMemory(DWORD size);
 DWORD kuhl_m_crypto_extractor_GetKeySize(DWORD bits);
+
+NTSTATUS kuhl_m_crypto_extract(int argc, wchar_t * argv[]);
 
 typedef struct _KIWI_CRYPTPROV {
 	PVOID CPAcquireContext;
@@ -604,3 +607,17 @@ typedef struct _KIWI_BCRYPT_HANDLE_KEY64 {
 #define RSAENH_KEY_64	0xe35a172cd96214a0
 #define DSSENH_KEY_64	0xa2491d83d96214a0
 #endif
+
+typedef struct _KIWI_CRYPT_SEARCH {
+	PKULL_M_MEMORY_HANDLE hMemory;
+	WORD Machine;
+	KIWI_CRYPTKEY32 ProcessKiwiCryptKey32;
+#ifdef _M_X64
+	KIWI_CRYPTKEY64 ProcessKiwiCryptKey64;
+#endif
+	BOOL bAllProcessKiwiCryptKey;
+	DWORD myPid;
+	DWORD prevPid;
+	DWORD currPid;
+	PCUNICODE_STRING processName;
+} KIWI_CRYPT_SEARCH, *PKIWI_CRYPT_SEARCH;
