+++ mimikatz/modules/crypto/kuhl_m_crypto_patch.h
@@ -1,0 +1,15 @@
+/*	Benjamin DELPY `gentilkiwi`
+	http://blog.gentilkiwi.com
+	benjamin@gentilkiwi.com
+	Licence : https://creativecommons.org/licenses/by/4.0/
+*/
+#pragma once
+#include "../kuhl_m.h"
+#include "../../../modules/kull_m_patch.h"
+#include "../../../modules/kull_m_crypto.h"
+
+typedef BOOL (WINAPI * PCP_EXPORTKEY) (IN HCRYPTPROV hProv, IN HCRYPTKEY hKey, IN HCRYPTKEY hPubKey, IN DWORD dwBlobType, IN DWORD dwFlags, OUT LPBYTE pbData, IN OUT LPDWORD pcbDataLen);
+PCP_EXPORTKEY K_CPExportKey;
+
+NTSTATUS kuhl_m_crypto_p_capi(int argc, wchar_t * argv[]);
+NTSTATUS kuhl_m_crypto_p_cng(int argc, wchar_t * argv[]);
